Add files via upload
Huber Timing is a website that offers comprehensive running results, including distance, time, and age, across the USA since 2012. The project involves a universal web scraper that efficiently collects and organizes this data using BeautifulSoup, pandas, matplotlib, and season libraries. The scraper ensures accurate and up-to-date information, making it a valuable resource for runners and enthusiasts.
Huber Timing:
https://www.hubertiming.com/results/2019Cinco5K